Spun out the QuestInfo class into it's own file. Also updated some typos, and added some more debug text. The bot actually runs now!
