+++ cogs/quest_handler.py
@@ -7,56 +7,14 @@ import traceback
 # webcolors is needed to take colour names and make them into a hex value
 # discord.py can work with
 import webcolors
-import json
 import db_handler as db
+from helpers import QuestInfo
 
 
 # -----------------------STATIC VARS----------------------
 # VS Code is annoying and runs the code in the home directory. This shouldn't be requred once I actually run the code as a standalone
 # Then I should just change this to "."
 FILE_LOCATION = "."
-
-
-# ---------------------HOLDER CLASSES---------------------
-class QuestInfo():
-    # This is just a blank class so I more easily can pass Quest data between
-    # objects
-    def __init__(
-            self,
-            quest_title: str,
-            contractor: str,
-            description: str,
-            reward: str,
-            embed_colour: str,
-            thread_id: int,
-            quest_role_id: int,
-            player_message: discord.Message,
-            pin_message_id: int,
-            players: str = None) -> None:
-        self.quest_title = quest_title
-        self.contractor = contractor
-        self.description = description
-        self.reward = reward
-        self.embed_colour = embed_colour
-        self.thread_id = thread_id
-        self.quest_role_id = quest_role_id
-        self.player_message = player_message
-        self.pin_message_id = pin_message_id
-        if players is not None:
-            self.players = json.loads(players)
-        else:
-            self.players = []
-
-    # adds a player (in the form of a member object) to the list
-    def add_player(self, member: discord.Member) -> None:
-        self.players.append(member)
-
-    # removes a player (as a member object) from the list
-    def remove_player(self, member: discord.Member) -> None:
-        self.players.remove(member)
-
-    def get_players_json(self) -> str:
-        return json.dumps(self._players)
 
 
 # ---------.----------PERSISTENT VIEWS--------------------
@@ -352,9 +310,9 @@ class DelQuest(discord.ui.Modal, title="Delete Quest"):
         if self.thread_del_flag.value.lower() == "yes":
             await interaction.guild.get_thread(self.quest_info.thread_id).delete()
         else:
+            # Lock the quest
             await interaction.guild.get_thread(self.quest_info.thread_id).edit(locked=True, archived=True)
-            print("locked")
-
+        
         #del role
         await interaction.guild.get_role(self.quest_info.quest_role_id).delete()
